+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Fix: minor change in handling of map of extra params.
+
 ## 1.0.0 (2022-06-07)
 
 Initial release of the commands library.